refactor: rename Model3D to ModelViewer
BREAKING CHANGE: Model3D is now named ModelViewer.

In cases where the GraphQL reference or fragment is mentioned, the name Model3D was maintained in order to match what the Storefront API calls it.
