+++ packages/hydrogen/src/components/MediaFile/README.md
@@ -2,7 +2,7 @@
 
 The `MediaFile` component renders the media for the Storefront API's
 [Media object](/api/storefront/reference/products/media). It renders an `Image`, a
-`Video`, an `ExternalVideo`, or a `Model3D` depending on the `mediaContentType` of the
+`Video`, an `ExternalVideo`, or a `ModelViewer` depending on the `mediaContentType` of the
 `media` provided as a prop.
 
 ## Example code
@@ -51,10 +51,10 @@ export function MyComponent() {
 
 ## Props
 
-| Name     | Type                                                                                         | Description                                                                     |
-| -------- | -------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------- |
-| media    | <code>MediaImageMedia &#124; Model3DMedia &#124; ExternalVideoMedia &#124; VideoMedia</code> | A [Media object](/api/storefront/reference/products/media).                     |
-| options? | <code>UndocumentedType &#124; UndocumentedType</code>                                        | The options for the `Image`, `Video`, `ExternalVideo`, or `Model3D` components. |
+| Name     | Type                                                                                             | Description                                                                         |
+| -------- | ------------------------------------------------------------------------------------------------ | ----------------------------------------------------------------------------------- |
+| media    | <code>MediaImageMedia &#124; ModelViewerMedia &#124; ExternalVideoMedia &#124; VideoMedia</code> | A [Media object](/api/storefront/reference/products/media).                         |
+| options? | <code>UndocumentedType &#124; UndocumentedType</code>                                            | The options for the `Image`, `Video`, `ExternalVideo`, or `ModelViewer` components. |
 
 ## Component type
