--- packages/hydrogen/src/components/Model3D/docs/component-type.md
@@ -1,3 +1,0 @@
-## Component type
-
-The `Model3D` component is a client component, which means that it renders on the client. For more information about component types, refer to [React Server Components](/custom-storefronts/hydrogen/framework/react-server-components).
