--- packages/hydrogen/src/components/Model3D/index.ts
@@ -1,1 +1,0 @@
-export {Model3D, Model3DProps} from './Model3D.client';
