+++ packages/hydrogen/src/components/ModelViewer/Model3D.graphql
