+++ packages/hydrogen/src/components/ModelViewer/Model3DFragment.ts
