+++ packages/hydrogen/src/components/ModelViewer/docs/component-type.md
@@ -1,0 +1,3 @@
+## Component type
+
+The `ModelViewer` component is a client component, which means that it renders on the client. For more information about component types, refer to [React Server Components](/custom-storefronts/hydrogen/framework/react-server-components).
