+++ packages/hydrogen/src/components/ModelViewer/index.ts
@@ -1,0 +1,1 @@
+export {ModelViewer, ModelViewerProps} from './ModelViewer.client';
