+++ packages/hydrogen/src/components/index.ts
@@ -11,8 +11,8 @@ export type {RawHtmlProps} from './RawHtml';
 export {RawHtml} from './RawHtml';
 export type {AddToCartButtonProps} from './AddToCartButton';
 export {AddToCartButton} from './AddToCartButton';
-export type {Model3DProps} from './Model3D';
-export {Model3D} from './Model3D';
+export type {ModelViewerProps} from './ModelViewer';
+export {ModelViewer} from './ModelViewer';
 export type {MoneyProps} from './Money';
 export {Money} from './Money';
 export {Metafield} from './Metafield';
